Refactor consumer module
