+++ src/backend/modules/consumer/src/queues/reconcileConsumerAuthClientOwnership.ts
@@ -1,6 +1,6 @@
 import { db, ID } from '@metorial/db';
 import { combineQueueProcessors, createQueue } from '@metorial/queue';
-import { consumerOAuthService } from '../services/consumerOAuth';
+import { consumerOAuthClientService } from '../services/consumerOAuth';
 
 let BATCH_SIZE = 100;
 
@@ -79,7 +79,7 @@ export let reconcileConsumerAuthClientOwnershipSingleQueueProcessor =
     if (consumerAuthClient.legacyDoNotUseConsumerSurface) {
       let consumerClient =
         consumerAuthClient.legacyDoNotUseConsumerClient ??
-        (await consumerOAuthService.upsertConsumerClient({
+        (await consumerOAuthClientService.upsertConsumerClient({
           consumerSurface: consumerAuthClient.legacyDoNotUseConsumerSurface,
           name: consumerAuthClient.name,
           redirectUris: consumerAuthClient.redirectUris
