+++ src/backend/modules/consumer/src/services/consumerAccess/index.ts
@@ -1,0 +1,5 @@
+export * from './accessPolicy';
+export * from './ares';
+export * from './consumerAccess';
+export * from './consumerAccessListing';
+export * from './consumerAccessRequest';
