+++ src/backend/modules/consumer/src/services/consumerEntities/index.ts
@@ -1,0 +1,5 @@
+export * from './consumerIntegration';
+export * from './consumerProviderCatalog';
+export * from './consumerProviderDeployment';
+export * from './consumerProviderSetupSession';
+export * from './consumerSkill';
