+++ src/backend/modules/consumer/src/services/consumerOAuth/index.ts
@@ -1,0 +1,7 @@
+export * from './_types';
+export * from './client';
+export * from './routing';
+export * from './registration';
+export * from './authorization';
+export * from './token';
+export * from './dashboard';
