+++ src/backend/modules/consumer/src/services/consumers/index.ts
@@ -1,0 +1,7 @@
+export * from './consumer';
+export * from './consumerAuth';
+export * from './consumerGroup';
+export * from './consumerInvite';
+export * from './consumerProfile';
+export * from './consumerSurface';
+export * from './consumerSurfaceProviderGroup';
