+++ src/backend/modules/consumer/src/services/index.ts
@@ -1,19 +1,5 @@
-export * from './accessPolicy';
-export * from './ares';
-export * from './consumer';
 export * from './consumerAccess';
-export * from './consumerAccessListing';
-export * from './consumerAccessRequest';
-export * from './consumerAuth';
-export * from './consumerGroup';
-export * from './consumerIntegration';
-export * from './consumerInvite';
+export * from './consumerEntities';
 export * from './consumerOAuth';
-export * from './consumerProfile';
-export * from './consumerProviderCatalog';
-export * from './consumerProviderDeployment';
-export * from './consumerProviderSetupSession';
-export * from './consumerSkill';
-export * from './consumerSurface';
-export * from './consumerSurfaceProviderGroup';
+export * from './consumers';
 export * from './portal';
