+++ src/backend/modules/consumer/test/consumerAccessRequestLifecycle.test.ts
@@ -56,7 +56,7 @@ describe('consumerAccessRequestUpdatedQueueProcessor', () => {
     vi.resetModules();
 
     db = (await import('@metorial/db')).db;
-    consumerAccessService = (await import('../src/services/consumerAccess'))
+    consumerAccessService = (await import('../src/services/consumerAccess/consumerAccess'))
       .consumerAccessService;
     indexConsumerAccessRequestSearchQueue = (
       await import('../src/queues/search/consumerAccessRequest')
